Get rid of Minify HTML option (PR #2813)
