+++ inc/3rd-party/hosting/wp-serveur.php
@@ -33,21 +33,3 @@ function rocket_wpserveur_varnish_field( $settings ) {
 add_filter( 'rocket_varnish_field_settings', 'rocket_wpserveur_varnish_field' );
 
 add_filter( 'rocket_display_input_varnish_auto_purge', '__return_false' );
-
-/**
- * Conflict with WP Serveur hosting: don't apply inline JS on all pages.
- *
- * @since 2.6.11
- *
- * @param array $html_options WP Rocket options array.
- *
- * @return array Updated WP Rocket options array.
- */
-function rocket_deactivate_inline_js_on_wp_serveur( $html_options ) {
-	if ( isset( $html_options['jsMinifier'] ) ) {
-		unset( $html_options['jsMinifier'] );
-	}
-
-	return $html_options;
-}
-add_action( 'rocket_minify_html_options', 'rocket_deactivate_inline_js_on_wp_serveur' );
