+++ inc/3rd-party/plugins/autoptimize.php
@@ -32,22 +32,6 @@ if ( class_exists( 'autoptimizeConfig' ) ) :
 	add_action( 'update_option_autoptimize_imgopt_settings', 'rocket_maybe_deactivate_lazyload', 10, 2 );
 
 	/**
-	 * Deactivate WP Rocket HTML Minification if Autoptimize HTML minification is enabled
-	 *
-	 * @since 2.9.5
-	 * @author Remy Perona
-	 *
-	 * @param string $old_value Previous autoptimize option value.
-	 * @param string $value New autoptimize option value.
-	 */
-	function rocket_maybe_deactivate_minify_html( $old_value, $value ) {
-		if ( $value !== $old_value && 'on' === $value ) {
-			update_rocket_option( 'minify_html', 0 );
-		}
-	}
-	add_action( 'update_option_autoptimize_html', 'rocket_maybe_deactivate_minify_html', 10, 2 );
-
-	/**
 	 * Deactivate WP Rocket CSS Minification if Autoptimize CSS minification is enabled
 	 *
 	 * @since 2.9.5
@@ -106,10 +90,6 @@ endif;
  * @author Remy Perona
  */
 function rocket_activate_autoptimize() {
-	if ( 'on' === get_option( 'autoptimize_html' ) ) {
-		update_rocket_option( 'minify_html', 0 );
-	}
-
 	if ( 'on' === get_option( 'autoptimize_css' ) ) {
 		update_rocket_option( 'minify_css', 0 );
 		update_rocket_option( 'minify_concatenate_css', 0 );
@@ -150,20 +130,6 @@ function rocket_maybe_disable_lazyload() {
 	}
 
 	return false;
-}
-
-/**
- * Disable WP Rocket HTML minification field if Autoptimize HTML minification is enabled
- *
- * @since 2.9.5
- * @author Remy Perona
- *
- * @return bool|null True if it is active
- */
-function rocket_maybe_disable_minify_html() {
-	if ( is_plugin_active( 'autoptimize/autoptimize.php' ) && 'on' === get_option( 'autoptimize_html' ) ) {
-		return true;
-	}
 }
 
 /**
