+++ inc/Engine/Admin/Deactivation/DeactivationIntent.php
@@ -150,7 +150,6 @@ mixpanel.init("a36067b00a263cce0299cfd960e26ecf", {
 			'minify_concatenate_css' => 0,
 			'minify_js'              => 0,
 			'minify_concatenate_js'  => 0,
-			'minify_html'            => 0,
 			'minify_google_fonts'    => 0,
 			'cdn'                    => 0,
 		];
