+++ inc/Engine/Admin/Settings/Page.php
@@ -582,10 +582,8 @@ class Page {
 		$this->settings->add_settings_sections(
 			[
 				'basic' => [
-					'title'  => __( 'Basic Settings', 'rocket' ),
-					'page'   => 'file_optimization',
-					// translators: %1$s = type of minification (HTML, CSS or JS), %2$s = “WP Rocket”.
-					'helper' => rocket_maybe_disable_minify_html() ? sprintf( __( '%1$s Minification is currently activated in <strong>Autoptimize</strong>. If you want to use %2$s’s minification, disable those options in Autoptimize.', 'rocket' ), 'HTML', WP_ROCKET_PLUGIN_NAME ) : '',
+					'title' => __( 'Basic Settings', 'rocket' ),
+					'page'  => 'file_optimization',
 				],
 				'css'   => [
 					'title'  => __( 'CSS Files', 'rocket' ),
@@ -612,21 +610,6 @@ class Page {
 
 		$this->settings->add_settings_fields(
 			[
-				'minify_html'            => [
-					'type'              => 'checkbox',
-					'label'             => __( 'Minify HTML', 'rocket' ),
-					'container_class'   => [
-						rocket_maybe_disable_minify_html() ? 'wpr-isDisabled' : '',
-					],
-					'description'       => __( 'Minifying HTML removes whitespace and comments to reduce the size.', 'rocket' ),
-					'section'           => 'basic',
-					'page'              => 'file_optimization',
-					'default'           => 0,
-					'sanitize_callback' => 'sanitize_checkbox',
-					'input_attr'        => [
-						'disabled' => rocket_maybe_disable_minify_html() ? 1 : 0,
-					],
-				],
 				'minify_google_fonts'    => [
 					'type'              => 'checkbox',
 					'label'             => __( 'Optimize Google Fonts', 'rocket' ),
