+++ inc/admin/ui/meta-boxes.php
@@ -71,7 +71,6 @@ function rocket_display_cache_options_meta_boxes() {
 			$fields = [
 				'lazyload'         => __( 'LazyLoad for images', 'rocket' ),
 				'lazyload_iframes' => __( 'LazyLoad for iframes/videos', 'rocket' ),
-				'minify_html'      => __( 'Minify HTML', 'rocket' ),
 				'minify_css'       => __( 'Minify/combine CSS', 'rocket' ),
 				'minify_js'        => __( 'Minify/combine JS', 'rocket' ),
 				'cdn'              => __( 'CDN', 'rocket' ),
@@ -153,7 +152,6 @@ function rocket_save_metabox_options() {
 		$fields = [
 			'lazyload',
 			'lazyload_iframes',
-			'minify_html',
 			'minify_css',
 			'minify_js',
 			'cdn',
