+++ inc/admin/ui/notices.php
@@ -155,7 +155,7 @@ function rocket_plugins_to_deactivate() {
 		$plugins['lazy-load-for-videos'] = 'lazy-load-for-videos/codeispoetry.php';
 	}
 
-	if ( get_rocket_option( 'minify_css' ) || get_rocket_option( 'minify_js' ) || get_rocket_option( 'minify_html' ) ) {
+	if ( get_rocket_option( 'minify_css' ) || get_rocket_option( 'minify_js' ) ) {
 		$plugins['wp-super-minify']         = 'wp-super-minify/wp-super-minify.php';
 		$plugins['bwp-minify']              = 'bwp-minify/bwp-minify.php';
 		$plugins['wp-minify']               = 'wp-minify/wp-minify.php';
@@ -168,11 +168,6 @@ function rocket_plugins_to_deactivate() {
 	if ( get_rocket_option( 'minify_css' ) || get_rocket_option( 'minify_js' ) ) {
 		$plugins['async-js-and-css']     = 'async-js-and-css/asyncJSandCSS.php';
 		$plugins['merge-minify-refresh'] = 'merge-minify-refresh/merge-minify-refresh.php';
-	}
-
-	if ( get_rocket_option( 'minify_html' ) ) {
-		$plugins['wp-html-compression'] = 'wp-html-compression/wp-html-compression.php';
-		$plugins['wp-compress-html']    = 'wp-compress-html/wp_compress_html.php';
 	}
 
 	if ( get_rocket_option( 'minify_js' ) ) {
