+++ tests/Unit/inc/Engine/Admin/Deactivation/DeactivationIntent/activateSafeMode.php
@@ -46,7 +46,6 @@ class Test_ActivateSafeMode extends TestCase {
 			'minify_concatenate_css' => 0,
 			'minify_js'              => 0,
 			'minify_concatenate_js'  => 0,
-			'minify_html'            => 0,
 			'minify_google_fonts'    => 0,
 			'cdn'                    => 0,
 		];
